+++ qoffer-domain/src/test/groovy/life/qbic/portal/portlet/customers/search/SearchCustomerSpec.groovy
@@ -3,10 +3,8 @@ package life.qbic.portal.portlet.customers.search
 import life.qbic.datamodel.dtos.business.AcademicTitle
 import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.Customer
-import life.qbic.portal.portlet.SearchCriteria
 import life.qbic.portal.portlet.exceptions.DatabaseQueryException
 import spock.lang.Specification
-
 
 /**
  * Test the search of customers