+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/customers/CustomerDatabaseQueries.groovy
@@ -45,10 +45,10 @@ class CustomerDatabaseQueries {
      */
     List<Customer> findPersonByName(String lastName){
         List<Customer> result = []
-        String query = "SELECT id, first_name, last_name, title, email from customer WHERE " +
+        String query = "SELECT id, first_name, last_name, title, email from person WHERE " +
             "last_name = ?"
 
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
 
         connection.withCloseable {
             def statement = it.prepareStatement(query)
@@ -97,10 +97,10 @@ class CustomerDatabaseQueries {
      */
     private List<Integer> getAffiliationIdsForPerson(int customerId) {
         List<Integer> result = []
-        String query = "SELECT affiliation_id FROM customer_affiliation WHERE " +
-            "customer_id = ?"
+        String query = "SELECT affiliation_id FROM person_affiliation WHERE " +
+            "person_id = ?"
 
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
 
         connection.withCloseable {
             def statement = it.prepareStatement(query)
@@ -124,7 +124,7 @@ class CustomerDatabaseQueries {
         String affiliationProperties = "organization, address_addition, street, postal_code, city, country, category"
         String query = "SELECT ${affiliationProperties} from affiliation WHERE " + "id = ?"
 
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
 
         connection.withCloseable {
             def statement = it.prepareStatement(query)
@@ -204,7 +204,7 @@ class CustomerDatabaseQueries {
         if (customerExists(customer)) {
             throw new DatabaseQueryException("Customer is already in the database.")
         }
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
         connection.setAutoCommit(false)
 
         connection.withCloseable {it ->
@@ -224,8 +224,8 @@ class CustomerDatabaseQueries {
     }
 
     private boolean customerExists(Customer customer) {
-        String query = "SELECT * FROM customer WHERE first_name = ? AND last_name = ? AND email = ?"
-        Connection connection = databaseSession.getConnection()
+        String query = "SELECT * FROM person WHERE first_name = ? AND last_name = ? AND email = ?"
+        Connection connection = databaseSession.connect()
 
         def customerAlreadyInDb = false
 
@@ -242,7 +242,7 @@ class CustomerDatabaseQueries {
     }
 
     private static int createNewCustomer(Connection connection, Customer customer) {
-        String query = "INSERT INTO customer (first_name, last_name, title, email) " +
+        String query = "INSERT INTO person (first_name, last_name, title, email) " +
             "VALUES(?, ?, ?, ?)"
 
         List<Integer> generatedKeys = []
@@ -263,7 +263,7 @@ class CustomerDatabaseQueries {
 
     private static void storeAffiliation(Connection connection, int customerId, List<Affiliation>
         affiliations) {
-        String query = "INSERT INTO customer_affiliation (affiliation_id, customer_id) " +
+        String query = "INSERT INTO person_affiliation (affiliation_id, customer_id) " +
             "VALUES(?, ?)"
 
         affiliations.each {affiliation ->
@@ -322,7 +322,7 @@ class CustomerDatabaseQueries {
         List<Affiliation> result = []
         String query = "SELECT * from affiliation"
 
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
 
         connection.withCloseable {
             def statement = it.prepareStatement(query)
@@ -352,7 +352,7 @@ class CustomerDatabaseQueries {
         if (affiliationExists(affiliation)) {
             throw new DatabaseQueryException("Affiliation is already in the database.")
         }
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
         connection.setAutoCommit(false)
 
         connection.withCloseable {it ->
@@ -387,7 +387,7 @@ class CustomerDatabaseQueries {
                 "AND city=? " +
                 "AND category=?"
 
-        Connection connection = databaseSession.getConnection()
+        Connection connection = databaseSession.connect()
 
         boolean affiliationAlreadyInDb = false